BSTModelKit.jl

BSTModelKit.jl is a package for constructing, solving, and analyzing Biochemical Systems Theory (BST) models of biochemical networks written in the Julia programming language. In particular, BSTModelKit.jl provides a set of tools for constructing and solving S-system BST representations of the form:

\[\begin{equation*} \frac{dX_{i}}{dt} = \alpha_{i}\prod_{j\in\mathcal{P}}X_{j}^{g_{ij}} - \beta_{i}\prod_{j\in\mathcal{R}}X_{j}^{h_{ij}}\qquad{i=1,\dots,n} \end{equation*}\]

where $X_{i}$ is the concentration of the species $i$, the values $\alpha_{i}$ and $\beta_{i}$ are kinetic parameters (rate constants), $\mathcal{P}$ is the set of species that produce $X_{i}$, $\mathcal{R}$ is the set of species that consume $X_{i}$, and $g_{ij}$ and $h_{ij}$ are the kinetic order coefficients relating species $i$ and $j$.

The S-system BST representation was developed by Savageau, Voit, and coworkers. For a nice introduction to the S-system representation, see:

Savageau M, Voit E, Irvine D. Biochemical systems theory and metabolic control theory: 1. fundamental similarities and differences. Math Biosci. 1987 86(2): 127-45. doi.org/10.1016/0025-5564(87)90007-1.

Savageau M. Biochemical systems analysis: a study of function and design in molecular biology. Reading, MA: Addison-Wesley; 1976.

Finally, BSTModelKit.jl is a research code, expect there will be (many x 1000) bugs, breaking changes (often), etc.

Installation and Requirements

BSTModelKit.jl can be installed, updated, or removed using the Julia package management system. To access the package management interface, open the Julia REPL, and start the package mode by pressing ]. While in package mode, to install BSTModelKit.jl, issue the command:

(@v1.1.x) pkg> add BSTModelKit

To use BSTModelKit.jl in your projects, issue the command:

julia> using BSTModelKit

Functions

BSTModelKit.sobol — Function

sobol(performance::Function, L::Array{Float64,1}, U::Array{Float64,1}; 
    number_of_samples::Int64 = 1000, orders::Array{Int64,1} = [0, 1, 2])

The sobol function is a wrapper around the gsa function that uses the Sobol method to perform a global sensitivity analysis.

Arguments

performance::Function: a function that takes a vector of parameters and returns a scalar performance metric.

L::Array{Float64,1}: a vector of lower bounds for the parameters.

U::Array{Float64,1}: a vector of upper bounds for the parameters.

number_of_samples::Int64: the number of samples to use in the analysis.

orders::Array{Int64,1}: the orders of sensitivity to compute.

Returns

a SobolResult object.

Base.indexin — Function

indexin(dd::Dict{String,Any},species_symbol::String; 
    key="total_species_list")::Union{Nothing,Int}

The indexin function is a utility function that returns the index of a species in the total species list of a model.

Arguments

dd::Dict{String,Any}: a dictionary that contains the model data.

species_symbol::String: the symbol of the species to find.

key::String: the key in the dictionary that contains the total species list.

Returns

an integer that represents the index of the species in the total species list, or nothing if the species is not found.

BSTModelKit.steadystate — Function

steadystate(model::BSTModel; tspan::Tuple{Float64,Float64} = (0.0,20.0), Δt::Float64 = 0.01, input::Union{Nothing,Function} = nothing) -> Array{Float64,1}

The steadystate function is used to evaluate the steady state of the model object that has been built using the build function.

Arguments

model::BSTModel: A model object that has been built using the build function.

tspan::Tuple{Float64,Float64}: A tuple that defines the time span for the simulation. The default is (0.0,20.0).

Δt::Float64: The time step for the simulation. The default is 0.01.

input::Union{Nothing,Function}: An optional input function that can be used to drive the simulation. The default is nothing.

Returns

A vector of state values that represent the steady state of the system.

BSTModelKit.evaluate — Function

evaluate(model::BSTModel; tspan::Tuple{Float64,Float64} = (0.0,20.0), Δt::Float64 = 0.01, 
    input::Union{Nothing,Function} = nothing) -> Tuple{Array{Float64,1}, Array{Float64,2}}

This function is used to evaluate the model object that has been built using the build function. The evaluate function will return a tuple with two elements: a vector of time points and a matrix of state values.

Arguments

model::BSTModel: A model object that has been built using the build function.

tspan::Tuple{Float64,Float64}: A tuple that defines the time span for the simulation. The default is (0.0,20.0).

Δt::Float64: The time step for the simulation. The default is 0.01.

input::Union{Nothing,Function}: An optional input function that can be used to drive the simulation. The default is nothing.

Returns

A tuple with two elements:

Array{Float64,1}: A vector of time points.

Array{Float64,2}: A matrix of state values.

BSTModelKit.BSTModel — Type

mutable struct BSTModel <: AbstractBSTModel

Mutable holding structure for the BST model object. This object is used to store all the data that is needed to simulate the model. The object is passed to the evaluate function to simulate the model. A BSTModel object is created using the build function.

Fields

number_of_dynamic_states::Int64: The number of dynamic states in the model.

number_of_static_states::Int64: The number of static states in the model.

list_of_dynamic_species::Array{String,1}: A list of the dynamic species in the model.

list_of_static_species::Array{String,1}: A list of the static species in the model.

list_of_reactions::Array{String,1}: A list of the reactions in the model.

total_species_list::Array{String,1}: A list of all the species in the model.

static_factors_array::Array{Float64,1}: A list of the static factors in the model.

initial_condition_array::Array{Float64,1}: A list of the initial conditions for the dynamic states.

S::Array{Float64,2}: The stoichiometry matrix.

G::Array{Float64,2}: The exponent matrix.

α::Array{Float64,1}: The rate constant vector.

Metadata fields

author::String: The author of the model.

version::String: The version of the model.

date::String: The date the model was created.

description::String: A description of the model.

BSTModelKit.build — Function

build(path::String) -> BSTModel

This function is used to build a BSTModel object from a model file. The model file can be in one of the following formats: bst, txt, jld2, dat, net, toml. The function will parse the file and build the model object.

Arguments

path::String: The path to the model file.

Returns

A BSTModel object. See the BSTModel documentation for more information on the fields of the model object.

BSTModelKit.morris — Function

morris(performance::Function, L::Array{Float64,1}, U::Array{Float64,1}; 
    number_of_samples::Int64 = 1000) -> Array{Float64,2}

The morris function is a wrapper around the gsa function that uses the Morris method to perform a global sensitivity analysis.

Arguments

performance::Function: a function that takes a vector of parameters and returns a scalar performance metric.

L::Array{Float64,1}: a vector of lower bounds for the parameters.

U::Array{Float64,1}: a vector of upper bounds for the parameters.

number_of_samples::Int64: the number of samples to use in the analysis.

Returns

a matrix of results where the first column is the mean and the second column is the variance of the sensitivity analysis.

Funding

The work described here was supported by the following grants:

The Interaction of Basal Risk, Pharmacological Ovulation Induction, Pregnancy and Delivery on Hemostatic Balance NIH NHLBI R-33 HL 141787 (PI’s Bernstein , Orfeo) and

The Pregnancy Phenotype and Predisposition to Preeclampsia NIH NHLBI R01 HL 71944 (PI Bernstein).

Disclaimer

This software is proved "AS IS" without warranty of any kind, express or implied, including but not limited to warranties of merchantability, fitness for a particular purpose, and non-infringement. In no event shall the authors or copyright holders be liable for any claim, damages, or other liability, whether in action or contract, tort or otherwise, arising from, out of, or in connection with the software or the use of other dealings in the software.